move ada stores into dedicated sub-folder
