+++ app/stores/ada/NetworkStatusStore.js
@@ -1,10 +1,10 @@
 // @flow
 import { observable, action, computed, runInAction } from 'mobx';
-import Store from './lib/Store';
-import Request from './lib/LocalizedRequest';
-import { ROUTES } from '../routes-config';
-import { Logger } from '../lib/logger';
-import type { GetSyncProgressResponse } from '../api/common';
+import Store from '../lib/Store';
+import Request from '../lib/LocalizedRequest';
+import { ROUTES } from '../../routes-config';
+import { Logger } from '../../lib/logger';
+import type { GetSyncProgressResponse } from '../../api/common';
 
 // To avoid slow reconnecting on store reset, we cache the most important props
 let cachedDifficulties = null;
