+++ app/stores/ada/NodeUpdateStore.js
@@ -1,11 +1,11 @@
 // @flow
 import { observable, action } from 'mobx';
-import Store from './lib/Store';
-import Request from './lib/LocalizedRequest';
+import Store from '../lib/Store';
+import Request from '../lib/LocalizedRequest';
 import type {
   ApplyUpdateResponse,
-} from '../api/ada/index';
-import type {NextUpdateResponse, PostponeUpdateResponse} from "../api/ada/index";
+} from '../../api/ada/index';
+import type {NextUpdateResponse, PostponeUpdateResponse} from "../../api/ada/index";
 
 export default class NodeUpdateStore extends Store {
