+++ app/stores/ada/WalletSettingsStore.js
@@ -1,14 +1,14 @@
 // @flow
 import { observable, action } from 'mobx';
 import _ from 'lodash';
-import Store from './lib/Store';
-import Request from './lib/LocalizedRequest';
-import globalMessages from '../i18n/global-messages';
+import Store from '../lib/Store';
+import Request from '../lib/LocalizedRequest';
+import globalMessages from '../../i18n/global-messages';
 import type {
   ExportWalletToFileResponse,
-} from '../api/ada/index';
-import type { WalletExportToFileParams } from '../actions/wallet-settings-actions';
-import type {UpdateWalletPasswordResponse, UpdateWalletResponse} from "../api/ada/index";
+} from '../../api/ada/index';
+import type { WalletExportToFileParams } from '../../actions/wallet-settings-actions';
+import type {UpdateWalletPasswordResponse, UpdateWalletResponse} from "../../api/ada/index";
 
 export default class WalletSettingsStore extends Store {
